requirements.txt
docx==0.2.4
matplotlib==3.2.2
numpy==1.21.4
opencv_python==4.4.0.46
Pillow==9.2.0
PyQt5==5.15.7
python_docx==0.8.11
pywin32==227
scikit_image==0.16.2
scipy==1.5.0
SimpleITK==2.1.1.2
skimage==0.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.47.0
